pyproject.toml
[tool.poetry]
name = "lojadelivros"
version = "0.1.0"
description = "lojadelivros app"
authors = ["Temistoclis <temistoclisxp58@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.3.3"
factory-boy = "^3.3.1"
django = "^5.1.2"
pytest-django = "^4.9.0"
djangorestframework = "^3.15.2"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
